updated fe deleteAllMaps

+++ src/App2Components/scenes/deleteAllMaps/index.jsx
@@ -4,11 +4,14 @@ import axios from "axios";
 
 function DeleteAllMaps() {
 
+    let userdetails=JSON.parse(window.sessionStorage.getItem("userdetails"));
+    const TenantId=userdetails.email;
+
     return (
         <div className="delete-all-maps-button">
             <h3>Delete all Maps</h3>
             <form onSubmit={(e) => {
-                axios.delete("http://localhost:5001/api/deleteAllMaps");
+                axios.delete(`http://localhost:5001/api/deleteAllMaps/${TenantId}`);
             }}>
                 <button type="submit">Delete All Maps</button>
             </form>
